Refactor httpBuildRequest and constructBaseUrl to accept BridgeStatus by ref and set it
