+++ source/extensions/filters/http/mcp_json_rest_bridge/BUILD
@@ -10,10 +10,20 @@ licenses(["notice"])  # Apache 2
 envoy_extension_package()
 
 envoy_cc_library(
+    name = "bridge_status_lib",
+    srcs = ["bridge_status.cc"],
+    hdrs = ["bridge_status.h"],
+    deps = [
+        "@abseil-cpp//absl/strings",
+    ],
+)
+
+envoy_cc_library(
     name = "mcp_json_rest_bridge_filter_lib",
     srcs = ["mcp_json_rest_bridge_filter.cc"],
     hdrs = ["mcp_json_rest_bridge_filter.h"],
     deps = [
+        ":bridge_status_lib",
         ":http_request_builder_lib",
         ":trace_context_lib",
         "//envoy/common:optref_lib",
@@ -48,6 +58,7 @@ envoy_cc_library(
     srcs = ["http_request_builder.cc"],
     hdrs = ["http_request_builder.h"],
     deps = [
+        ":bridge_status_lib",
         "//source/common/common:regex_lib",
         "//source/common/common:utility_lib",
         "//source/common/http:utility_lib",
