+++ source/extensions/filters/http/mcp_json_rest_bridge/http_request_builder.cc
@@ -181,12 +181,14 @@ void substitutePathTemplateVariable(std::string& url, absl::string_view element,
 
 absl::StatusOr<std::string> constructBaseUrl(absl::string_view pattern,
                                              const absl::flat_hash_set<std::string>& templates,
-                                             const nlohmann::json& arguments) {
+                                             const nlohmann::json& arguments,
+                                             BridgeStatus& bridge_status) {
   std::string base_url = std::string(pattern);
   for (const auto& element : templates) {
     absl::StatusOr<nlohmann::json> template_value_json = getJsonValue(arguments, element);
     if (!template_value_json.ok()) {
-      return template_value_json.status();
+      bridge_status = BridgeStatus::RequestToolsCallMissingRequiredArg;
+      return absl::InvalidArgumentError("Missing required argument");
     }
     const std::string raw_value = jsonValueToString(*template_value_json);
 
@@ -196,8 +198,8 @@ absl::StatusOr<std::string> constructBaseUrl(absl::string_view pattern,
     // traversal, so treat it as a separator here too.
     for (const absl::string_view segment : absl::StrSplit(raw_value, absl::ByAnyChar("\\/"))) {
       if (segment == "." || segment == "..") {
-        return absl::PermissionDeniedError(absl::StrCat(
-            "path template variable '", element, "' must not contain path traversal segments"));
+        bridge_status = BridgeStatus::RequestToolsCallPathTraversalRejected;
+        return absl::InvalidArgumentError("Path traversal rejected");
       }
     }
     // A simple variable (`{id}`) fills one segment, so its '/' is encoded; a variable with an
@@ -216,7 +218,7 @@ absl::StatusOr<std::string> constructBaseUrl(absl::string_view pattern,
 
 absl::StatusOr<HttpRequest> buildHttpRequest(
     const envoy::extensions::filters::http::mcp_json_rest_bridge::v3::HttpRule& http_rule,
-    const nlohmann::json& arguments) {
+    const nlohmann::json& arguments, BridgeStatus& bridge_status) {
   std::string pattern;
   std::string method;
   // TODO(guoyilin42): Add validation to ensure exactly one HTTP method is specified.
@@ -236,7 +238,8 @@ absl::StatusOr<HttpRequest> buildHttpRequest(
     method = "PATCH";
     pattern = http_rule.patch();
   } else {
-    return absl::InternalError("HttpRule is malformed");
+    bridge_status = BridgeStatus::InternalToolsCallInvalidHttpRule;
+    return absl::InvalidArgumentError("HttpRule is malformed");
   }
   absl::string_view url_template = pattern;
   absl::flat_hash_set<std::string> templates;
@@ -245,7 +248,7 @@ absl::StatusOr<HttpRequest> buildHttpRequest(
   while (RE2::FindAndConsume(&url_template, *template_regex, &template_capture)) {
     templates.insert(template_capture);
   }
-  absl::StatusOr<std::string> url = constructBaseUrl(pattern, templates, arguments);
+  absl::StatusOr<std::string> url = constructBaseUrl(pattern, templates, arguments, bridge_status);
   if (!url.ok()) {
     return url.status();
   }
@@ -256,14 +259,16 @@ absl::StatusOr<HttpRequest> buildHttpRequest(
     if (auto status =
             constructQueryParams(query_params, http_rule.body(), arguments, templates, base_path);
         !status.ok()) {
-      return status;
+      bridge_status = BridgeStatus::RequestToolsCallMissingRequiredArg;
+      return absl::InvalidArgumentError("Missing required argument");
     }
   }
   appendQueryParamsToBaseUrl(*url, query_params);
 
   absl::StatusOr<json> http_body = constructRequestBody(http_rule.body(), templates, arguments);
   if (!http_body.ok()) {
-    return http_body.status();
+    bridge_status = BridgeStatus::RequestToolsCallMissingRequiredArg;
+    return absl::InvalidArgumentError("Missing required argument");
   }
 
   return HttpRequest{
