+++ source/extensions/filters/http/mcp_json_rest_bridge/mcp_json_rest_bridge_filter.cc
@@ -177,52 +177,6 @@ std::array<EndpointKey, 4> getEndpointLookupKeys(absl::string_view host, absl::s
 
 } // namespace
 
-absl::string_view bridgeStatusToString(BridgeStatus status) {
-  switch (status) {
-  case BridgeStatus::Ok:
-    return BridgeStatusValues::OK;
-  case BridgeStatus::HttpRequestMethodNotPost:
-    return BridgeStatusValues::HTTP_REQUEST_METHOD_NOT_POST;
-  case BridgeStatus::RequestTooLarge:
-    return BridgeStatusValues::REQUEST_TOO_LARGE;
-  case BridgeStatus::RequestFailedToParseJsonRpc:
-    return BridgeStatusValues::REQUEST_FAILED_TO_PARSE_JSON_RPC;
-  case BridgeStatus::RequestUnsupportedMcpVersion:
-    return BridgeStatusValues::REQUEST_UNSUPPORTED_MCP_VERSION;
-  case BridgeStatus::RequestInitializeNotValid:
-    return BridgeStatusValues::REQUEST_INITIALIZE_NOT_VALID;
-  case BridgeStatus::RequestMcpMethodNotSupported:
-    return BridgeStatusValues::REQUEST_MCP_METHOD_NOT_SUPPORTED;
-  case BridgeStatus::RequestMcpMethodMalformed:
-    return BridgeStatusValues::REQUEST_MCP_METHOD_MALFORMED;
-  case BridgeStatus::RequestIdNotFound:
-    return BridgeStatusValues::REQUEST_ID_NOT_FOUND;
-  case BridgeStatus::RequestToolsCallToolNameMissing:
-    return BridgeStatusValues::REQUEST_TOOLS_CALL_TOOL_NAME_MISSING;
-  case BridgeStatus::RequestToolsCallToolNameUnknown:
-    return BridgeStatusValues::REQUEST_TOOLS_CALL_TOOL_NAME_UNKNOWN;
-  case BridgeStatus::RequestToolsCallArgumentsMalformed:
-    return BridgeStatusValues::REQUEST_TOOLS_CALL_ARGUMENTS_MALFORMED;
-  case BridgeStatus::RequestToolsCallMissingRequiredArg:
-    return BridgeStatusValues::REQUEST_TOOLS_CALL_MISSING_REQUIRED_ARG;
-  case BridgeStatus::RequestToolsCallPathTraversalRejected:
-    return BridgeStatusValues::REQUEST_TOOLS_CALL_PATH_TRAVERSAL_REJECTED;
-  case BridgeStatus::InternalToolsCallInvalidHttpRule:
-    return BridgeStatusValues::INTERNAL_TOOLS_CALL_INVALID_HTTP_RULE;
-  case BridgeStatus::InternalToolsListPassthrough:
-    return BridgeStatusValues::INTERNAL_TOOLS_LIST_PASSTHROUGH;
-  case BridgeStatus::ResponseTooLarge:
-    return BridgeStatusValues::RESPONSE_TOO_LARGE;
-  case BridgeStatus::ResponseToolsCallInvalidUtf8:
-    return BridgeStatusValues::RESPONSE_TOOLS_CALL_INVALID_UTF8;
-  case BridgeStatus::ResponseHttpStatusError:
-    return BridgeStatusValues::RESPONSE_HTTP_STATUS_ERROR;
-  case BridgeStatus::ResponseFailedToParseJsonRpc:
-    return BridgeStatusValues::RESPONSE_FAILED_TO_PARSE_JSON_RPC;
-  }
-  return "UNKNOWN";
-}
-
 McpJsonRestBridgeFilterConfig::McpJsonRestBridgeFilterConfig(
     const envoy::extensions::filters::http::mcp_json_rest_bridge::v3::McpJsonRestBridge&
         proto_config)
@@ -1154,30 +1108,14 @@ void McpJsonRestBridgeFilter::mapMcpToolToApiBackend(
   const nlohmann::json empty_arguments = nlohmann::json::object();
   const nlohmann::json& arguments = arguments_it != params.end() ? *arguments_it : empty_arguments;
 
-  absl::StatusOr<HttpRequest> http_request = buildHttpRequest(*http_rule, arguments);
+  BridgeStatus bridge_status = BridgeStatus::Ok;
+  absl::StatusOr<HttpRequest> http_request = buildHttpRequest(*http_rule, arguments, bridge_status);
   if (!http_request.ok()) {
     ENVOY_STREAM_LOG(error, "Failed to build HTTP request for method: {} with status: {}",
                      *decoder_callbacks_, tool_name, http_request.status());
-    BridgeStatus bridge_status;
-    std::string error_msg;
-    switch (http_request.status().code()) {
-    case absl::StatusCode::kPermissionDenied:
-      bridge_status = BridgeStatus::RequestToolsCallPathTraversalRejected;
-      error_msg = "Path traversal rejected";
-      break;
-    case absl::StatusCode::kInternal:
-      bridge_status = BridgeStatus::InternalToolsCallInvalidHttpRule;
-      error_msg = "HttpRule is malformed";
-      break;
-    case absl::StatusCode::kInvalidArgument:
-    default:
-      bridge_status = BridgeStatus::RequestToolsCallMissingRequiredArg;
-      error_msg = "Missing required argument";
-      break;
-    }
     sendErrorResponse(Http::Code::OK, bridge_status,
-                      generateErrorJsonResponse(-32602, error_msg).dump(), nullptr,
-                      McpConstants::Methods::TOOLS_CALL, params);
+                      generateErrorJsonResponse(-32602, http_request.status().message()).dump(),
+                      nullptr, McpConstants::Methods::TOOLS_CALL, params);
     return;
   }
